% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhittakerPlot.R
\name{WhittakerPlot}
\alias{WhittakerPlot}
\title{Whittaker Plot}
\usage{
WhittakerPlot(data, sites, ylog = FALSE, maxRank = nrow(data))
}
\arguments{
\item{data}{A data frame containing the species abundance data. The first column should represent the taxon, and the rest of the columns represent the sites.}

\item{sites}{A vector of site names to be included in the plot.}

\item{ylog}{A logical value indicating whether the y-axis should be on a log scale or not. The default value is FALSE.}

\item{maxRank}{An optional numeric value specifying the maximum rank for the x-axis. If not specified, the function will use the number of rows in the data.}
}
\value{
A Whittaker Plot is displayed.
}
\description{
This function generates a Whittaker Plot, which is a type of rank-abundance plot used in ecology to display relative species abundance, the number of species, and species richness in a community.
}
\examples{
# Example data
transposed_data <- data.frame(
Species = c("Species1", "Species2", "Species3", "Species4", "Species5"),
Site1 = c(5, 3, 0, 7, 2),
Site2 = c(2, 0, 6, 1, 4),
Site3 = c(1, 8, 2, 0, 5),
Site4 = c(7, 1, 3, 4, 6),
Site5 = c(3, 2, 5, 6, 1)
)
# Use the WhittakerPlot function
WhittakerPlot(transposed_data, c("Site1", "Site5"))
}
